Move valid input method to a new file
