+++ src/modules/valid_input.rb
@@ -1,0 +1,10 @@
+def get_user_input(prompt, valid_responses)
+  while true
+    print prompt
+    input = gets.chomp
+    break if valid_responses.include?(input)
+
+    puts "\n🛑 ❌ Invalid input. Please enter one of the following: #{valid_responses.join(', ')} \n".red
+  end
+  input
+end
